# perf: replace _uniform method to remove iteration on tensor