@vfdev-5 2023-08-23
Nice improvements, @MarcBresson !

For equality tests I would use `torch.testing.assert_close(old_uniform(kernel_size), new_uniform(kernel_size))`.

I wonder how to see that 
>    start_uniform_index = max(kernel_size // 2 - 2, 0)
>    end_uniform_index = min(kernel_size // 2 + 3, kernel_size)
>    kernel[start_uniform_index:end_uniform_index] = 1 / (max_ - min_)

is equivalent to 
```
        if min <= j <= max:
            kernel[i] = 1 / (max - min)
```
?