@MarcBresson 2023-08-24
It was hard to wrap my head around the possible decisions that led to this code.

Basically, the former code was creating a tensor that went from `-kernel_size // 2` to `kernel_size // 2`. Then all the values of this tensor that were < to -2.5 or > to 2.5 were replaced by 0, and the other were set to `1 / (max - min)`.

The new code just put 0 everywhere then compute the indices that should be set to `1 / (max - min)`.

The only difference between the two codes is when the kernel size is an even number, but this raise an error beforehand.